refactor: Accordion - review 반영
